requirements.txt
kaleido==0.2.1
lxml==5.2.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pipenv==2024.4.0
plotly==5.24.1
scikit-learn==1.5.2
screeninfo==0.8.1
stqdm==0.0.5
streamlit==1.40.1
streamlit-extras==0.5.0
streamlit-image-select==0.6.0
tables==3.10.1
toml==0.10.2
tqdm==4.67.0
pytest==8.3.3


